requirements.txt
paho-mqtt
netifaces
